third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-lifetime-mediakeysession-reference.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test MediaKeySession lifetime</title>
        <script src="../w3c-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
        <script src="../../resources/gc.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script>
            // Since MediaKeySession (but not MediaKeys) are ActiveDOMObjects,
            // we can determine when they are garbage collected.
            // MediaKeySessions remain as long as:
            //   JavaScript has a reference to it
            //   OR (MediaKeys is around AND the session has not received a close() event)
            test(function()
            {
                var initData = new Uint8Array([0, 1, 2, 3, 4, 5, 6, 7]);
                var startingActiveDOMObjectCount = window.internals.activeDOMObjectCount(document);

                function numActiveDOMObjectsCreated()
                {
                    return window.internals.activeDOMObjectCount(document) - startingActiveDOMObjectCount;
                }

                var mediaKeys = new MediaKeys("org.w3.clearkey");
                assert_equals(mediaKeys.keySystem, 'org.w3.clearkey');

                // Verify MediaKeys are not ActiveDOMObjects.
                assert_equals(numActiveDOMObjectsCreated(), 0, "MediaKeys created an ActiveDOMObject");

                // Create 3 sessions.
                var mediaKeySession1 = mediaKeys.createSession("video/webm", initData);
                assert_equals(numActiveDOMObjectsCreated(), 1);
                var mediaKeySession2 = mediaKeys.createSession("video/webm", initData);
                assert_equals(numActiveDOMObjectsCreated(), 2);
                var mediaKeySession3 = mediaKeys.createSession("video/webm", initData);
                assert_equals(numActiveDOMObjectsCreated(), 3);

                // Run gc(). All sessions should remain as we have a reference
                // to each one.
                gc();
                assert_equals(numActiveDOMObjectsCreated(), 3);

                // Now drop references to 2 of the sessions. Even though we
                // don't have a reference, MediaKeys is still around (and the
                // sessions aren't closed), so the objects won't be collected.
                mediaKeySession2 = null;
                mediaKeySession3 = null;
                gc();
                assert_equals(numActiveDOMObjectsCreated(), 3);

                // Now drop the reference to MediaKeys. It and the 2
                // unreferenced sessions should be collected. Since
                // MediaKeySessions remain alive as long as MediaKeys is around,
                // it is possible that gc() checks one or both MediaKeySession
                // objects first, and doesn't collect them since MediaKeys
                // hasn't been collected yet. Thus run gc() twice, to ensure
                // that the unreferenced MediaKeySession objects get collected.
                mediaKeys = null;
                gc();
                gc();
                assert_equals(numActiveDOMObjectsCreated(), 1);

                // Drop the reference to the first session. It should get
                // collected now since MediaKeys is gone.
                mediaKeySession1 = null;
                gc();
                assert_equals(numActiveDOMObjectsCreated(), 0);
            }, "MediaKeySession lifetime without release()");
        </script>
    </body>
</html>
